% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XChiSquare.R
\name{XChiSquare}
\alias{XChiSquare}
\title{The (non-central) Chi-Squared Distribution}
\usage{
XChiSquare(df = 2, ConfidenceLevel = 0.95)
}
\arguments{
\item{df}{}

\item{ConfidenceLevel}{}
}
\value{

}
\description{
The mean on ChiSquare is = Degrees Of Freedom
SD is sqrt(2*DF)
}
