% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XDescribeRandomVariable.R
\name{XDescribeRandomVariable}
\alias{XDescribeRandomVariable}
\title{Describe Numerical Random Variable}
\usage{
XDescribeRandomVariable(X, p)
}
\arguments{
\item{X}{Vector of Numerical Values}

\item{p}{Associated Vector of probabilities (Optional)}
}
\description{
Describe Numerical Random Variable
}
\examples{
XDescribeRandomVariable(1:10,rep(1/10,10))
}
