% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XElectricForce.R
\name{XElectricForce}
\alias{XElectricForce}
\title{Electric Force}
\usage{
XElectricForce(...)
}
\arguments{
\item{...}{A List of Points (x,y,z,Q)}
}
\value{

}
\description{
Calculate the Resulting Force applyed at 1st Point caused by the other Points
using Coulombs Law: F = ke Q1 Q2 /R^2
}
