% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XExecSQL.R
\name{XExecSQL}
\alias{XExecSQL}
\title{Execute SQL query and return data frame}
\usage{
XExecSQL(subsys, query)
}
\arguments{
\item{subsys}{subsystem}

\item{sql}{query}
}
\description{
Execute SQL query and return data frame
}
