% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XFillMissingNumbers.R
\name{XFillMissingNumbers}
\alias{XFillMissingNumbers}
\title{Fill missing Numbers}
\usage{
XFillMissingNumbers(df, formula)
}
\arguments{
\item{df}{input dataframe}

\item{formula}{formula to predict missing values}
}
\description{
Fill missing Numbers
}
