% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XFunctions.R
\name{XFunctions}
\alias{XFunctions}
\title{Receive Package List, returns Function names}
\usage{
XFunctions(...)
}
\arguments{
\item{...}{1 or more Packages comma separated or in vector format}
}
\value{

}
\description{
Loads a Data Frame in the Global Environment named: Loaded.Functions
}
