% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XHistogram.R
\name{XHistogram}
\alias{XHistogram}
\title{Plot Histogram with superimposed best-fitting normal distribution}
\usage{
XHistogram(
  distribution,
  probability = FALSE,
  x1 = mean - sd,
  x2 = mean + sd,
  name = "A",
  color = "green"
)
}
\arguments{
\item{distribution}{Vector of values for the histogram}

\item{probability}{Optional, Boolean. If TRUE, plot histogram of Probabilities ( default is False )}

\item{x1}{Optional, A Special point 1 to study ( default is x1=mean-sd )}

\item{x2}{Optional, A Special point 2 to study ( default is x2=mean+sd)}

\item{name}{Optional, A Name for the Data ( default is the name of the distribution )}

\item{color}{Optional, A Color for the histogram ( defualt is green )}
}
\value{

}
\description{
Can Plot histogram of Counts or Histogram of Probabilities
}
\examples{
XHistogram(iris$Sepal.Length)
XHistogram(iris$Sepal.Length,probability = T,x1=6,x2=7,name="Sepal Lenth",color="yellow")
}
