% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XHistogramDiscrete.R
\name{XHistogramDiscrete}
\alias{XHistogramDiscrete}
\title{Plot Histogram for a Vector of Integers}
\usage{
XHistogramDiscrete(v)
}
\arguments{
\item{v}{A vector of Values}
}
\value{

}
\description{
It plots an Histogram with one bar for each distinct value
Try XHistogramDiscrete(mtcars$cyl) and compare it with hist(mtcars$cyl)
}
\examples{
XHistogramDiscrete(c(-3,5,5,4,10,8,8,NA))
XHistogramDiscrete(mtcars$cyl) 
}
