% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XIsNormalyDistributed.R
\name{XIsNormalyDistributed}
\alias{XIsNormalyDistributed}
\title{Verify if a Distribution is normally distributed}
\usage{
XIsNormalyDistributed(X)
}
\arguments{
\item{X}{A Numeric Vector}
}
\value{

}
\description{
Verify if a Distribution is normally distributed
}
\examples{
XIsNormalyDistributed(iris$Sepal.Length)
}
