% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XJointCategorical.R
\name{XJointCategorical}
\alias{XJointCategorical}
\title{Joint Distribution between 2 Categorical Random Variables X,Y with probability table p}
\usage{
XJointCategorical(X, Y, p)
}
\arguments{
\item{X}{Random Variable X}

\item{Y}{Random Variable Y}

\item{p}{Totals Table or Probability Table}
}
\description{
Joint Distribution between 2 Categorical Random Variables X,Y with probability table p
}
