% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XJointNumerical.R
\name{XJointNumerical}
\alias{XJointNumerical}
\title{Joint Distribution between 2 Numerical Random Variables X,Y with probability table p
Also Analyse the Portfolio Combination of a\emph{X + b}Y}
\usage{
XJointNumerical(X, Y, p, a = 1, b = 1)
}
\arguments{
\item{X}{Random Variable X}

\item{Y}{Random Variable Y}

\item{p}{Totals Table or Probability Table}

\item{a}{Portfolio Combination of a\emph{X + b}Y}

\item{b}{Portfolio Combination of a\emph{X + b}Y}
}
\description{
Joint Distribution between 2 Numerical Random Variables X,Y with probability table p
Also Analyse the Portfolio Combination of a\emph{X + b}Y
}
