% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XLibrary.R
\name{XLibrary}
\alias{XLibrary}
\title{Automatically install/load Packages}
\usage{
XLibrary(...)
}
\arguments{
\item{package}{0, 1 or more comma separated packages}
}
\description{
Receives a list of packages to Install/Load \cr
Adds the received packages to a list of default packages to allways load \cr
Installs missing packages \cr
Load Packages \cr
Inspect Packages for datasets \cr
Creates 3 data frames on Global Environment: \cr
Loaded.Packages \cr
Loaded.Datasets \cr
Loaded.Vignettes
}
\examples{
XLibrary()
XLibrary("benchmarkme")
XLibrary("package1","package2")
}
\keyword{XLibrary}
