% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XListEnvironments.R
\name{XListEnvironments}
\alias{XListEnvironments}
\title{List chain of Environments, starting on a given Environment}
\usage{
XListEnvironments(env = globalenv())
}
\arguments{
\item{env}{Receives an environment, defaults for globalenv()}
}
\value{

}
\description{
List chain of Environments, starting on a given Environment
}
