% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMode.R
\name{XMode}
\alias{XMode}
\title{Returns the statistical mode of a vector Unimodal or MultiModal}
\usage{
XMode(x)
}
\arguments{
\item{x}{A numerical vector}
}
\value{

}
\description{
Returns the statistical mode of a vector Unimodal or MultiModal
}
\examples{
XMode()
XMode(mtcars$cyl)
XMode(c(2,1,2,3,1,2,3,4,1,5,5,3,2,3))
XMode(c("o","it","the","it","o","it","o"))
}
