% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPlot.R
\name{XPlot}
\alias{XPlot}
\title{Plots a function}
\usage{
XPlot(f, x1, x2, x0 = (x1 + x2)/2)
}
\arguments{
\item{x1}{from x1}

\item{x2}{to x2}

\item{x0}{Optional, studies the point x0, defaults to x0=(x1+x2)/2}
}
\description{
Plots a function
}
\examples{
XPlot(function(x) 2*x^2+3*x,-5,5,1)
XPlot(function(x) sin(x),0,2*pi)
}
