% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPlotBinomial.R
\name{XPlotBinomial}
\alias{XPlotBinomial}
\title{Plots The Binomial Distribution B(n,p)}
\usage{
XPlotBinomial(n, p)
}
\arguments{
\item{n}{Integer parameter n}

\item{p}{Probability p between 0 and 1}
}
\description{
Plots The Binomial Distribution B(n,p)
}
\examples{
XPlotBinomial(10,0.3)
}
