% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPlotGeometric.R
\name{XPlotGeometric}
\alias{XPlotGeometric}
\title{Plots The Geometric Distribution G(p)}
\usage{
XPlotGeometric(p)
}
\arguments{
\item{p}{Probability p between 0 and 1}
}
\description{
Plots The Geometric Distribution G(p)
}
\examples{
XPlotGeometric(0.3)
}
