% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPlotMulti.R
\name{XPlotMulti}
\alias{XPlotMulti}
\title{Plot a List of Functions}
\usage{
XPlotMulti(myfunctions, x1, x2)
}
\arguments{
\item{myfunctions}{Receives a Vector of 1 or more Functions}

\item{x1}{plot from}

\item{x2}{plot to}
}
\value{

}
\description{
Plot a List of Functions
}
\examples{
f1 = function(x) cos(x)
f2 = function(x) sin(x)
f3 = function(x) x^2
XPlotMulti(c(f1,f2,f3),-pi,pi)
}
