% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPlotNegativeBinomial.R
\name{XPlotNegativeBinomial}
\alias{XPlotNegativeBinomial}
\title{Plots The Negative Binomial Distribution NBinom(r,p)}
\usage{
XPlotNegativeBinomial(r, p)
}
\arguments{
\item{r}{Number of Expected Successes}

\item{p}{Bernoulli Probability p between 0 and 1}
}
\description{
Plots The Negative Binomial Distribution NBinom(r,p)
}
\examples{
XPlotNegativeBinomial(3,1/6)
}
