% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPlotOjive.R
\name{XPlotOjive}
\alias{XPlotOjive}
\title{Plots The Cumulative Frequency Graph ( Ojive )}
\usage{
XPlotOjive(d)
}
\arguments{
\item{d}{Distribution Vector}
}
\description{
Plots The Cumulative Frequency Graph ( Ojive )
}
\examples{
XPlotOjive((iris$Sepal.Length))
}
