% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XRandomVariable.R
\name{XRandomVariable}
\alias{XRandomVariable}
\title{Discrete Random Variable
Generic Distribution of n trials with replacement from a box
The Analytical Expression that gives the probability of X is the Density Function
The Integral of the Density Function is the CDF}
\usage{
XRandomVariable(X, n = 1, p)
}
\arguments{
\item{X}{Random Variable Vector}

\item{n}{Number of Trials}

\item{p}{Probability Vector ( optional )}
}
\value{

}
\description{
PMF (Probability Mass Function) uses discrete random variables. 
PDF (Probability Density Function) uses continuous random variables. 
Based on studies, PDF is the derivative of CDF, which is the cumulative distribution function. 
CDF is used to determine the probability wherein a continuous random variable would occur within any measurable subset of a certain range
}
\details{
Sums and Differences of Random Variables: Effect on the Mean
E(X + Y) = E(X) + E(Y) 

E(X - Y) = E(X) - E(Y) 
Sums and Differences of Independent Random Variables: Effect on Variance
Var(X + Y) = Var(X - Y) = Var(X) + Var(Y)
}
\examples{
XRandomVariable(1:6,n=1)
}
