% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XRandomVariableContinuous.R
\name{XRandomVariableContinuous}
\alias{XRandomVariableContinuous}
\title{Analyse and Plot Continuous Random Variable}
\usage{
XRandomVariableContinuous(f, x1, x2)
}
\arguments{
\item{f}{Probability Function}

\item{x1}{Defined from x1}

\item{x2}{to x2}
}
\description{
Analyse and Plot Continuous Random Variable
}
\examples{
XRandomVariableContinuous(f = function(x) 3*x*(3-x)/10,x1=0,x2=2)
}
