% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XRandomVariableDiscrete.R
\name{XRandomVariableDiscrete}
\alias{XRandomVariableDiscrete}
\title{Analyse and Plot Discrete Random Variable}
\usage{
XRandomVariableDiscrete(X, p)
}
\arguments{
\item{X}{Vector of Numerical Values}

\item{p}{Associated Vector of probabilities (Optional)}
}
\description{
Analyse and Plot Discrete Random Variable
}
\examples{
XRandomVariableDiscrete(1:10)
}
