% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XScatterPlot.R
\name{XScatterPlot}
\alias{XScatterPlot}
\title{Scatter Plot for 2 continuous variables}
\usage{
XScatterPlot(x, y)
}
\arguments{
\item{x}{x data}

\item{y}{y data}
}
\value{

}
\description{
Scatter Plot for 2 continuous variables
}
\examples{
XScatterPlot(mtcars$wt, mtcars$mpg)
}
