% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XScatterPlot2.R
\name{XScatterPlot2}
\alias{XScatterPlot2}
\title{Compute best fit line and make prediction}
\usage{
XScatterPlot2(xMean, xSD, yMean, ySD, r, x)
}
\arguments{
\item{xMean}{Mean from x vector}

\item{xSD}{SD from x vector}

\item{yMean}{Mean from y vector}

\item{ySD}{SD from y vector}

\item{r}{Correlation Factor}

\item{x}{Predict for this x}
}
\description{
Compute best fit line and make prediction
}
\examples{
XScatterPlot2(36.92,37.79,24.71,25.21,0.978,35)
}
