% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XSplit.R
\name{XSplit}
\alias{XSplit}
\title{Split a Data Frame in Training and Test DF ( only valid for categorical target )}
\usage{
XSplit(data, p, target)
}
\arguments{
\item{data}{Input Data Frame}

\item{p}{proportion for training}

\item{target}{name of the target variable}
}
\description{
Split a Data Frame in Training and Test DF ( only valid for categorical target )
}
