% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XStatsCIproportion2.R
\name{XStatsCIproportion2}
\alias{XStatsCIproportion2}
\title{Estimates Confidence Interval for the difference between 2 Sample Proportions}
\usage{
XStatsCIproportion2(n1, p1, n2, p2, CL = 0.95, SL = 0.05)
}
\arguments{
\item{n1}{Sample 1 Size}

\item{p1}{Sample 1 Proportion}

\item{n2}{Sample 2 Size}

\item{p2}{Sample 2 Proportion}

\item{CL}{Confidence Level ( default=0.95 )}

\item{SL}{Significance Level for the Hypothesis test for difference in proportions ( default=0.05 )}
}
\description{
Inferences on one proportion are:
\cr Confidence Intervals
\cr Significance Tests
\cr
\cr Criteria for both, are:
\cr Randomness
\cr Normality
\cr Independence
}
\examples{
XStatsCIproportion2(120,54/120,140,77/140,0.9)
}
