% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XStatsOneSampleTtest.R
\name{XStatsOneSampleTtest}
\alias{XStatsOneSampleTtest}
\title{Estimates Population Mean from a Simple Random Sample (SRS)}
\usage{
XStatsOneSampleTtest(n, m, S, CL = 0.95, N = 1e+06, t)
}
\arguments{
\item{n}{Sample Size}

\item{m}{Sample mean ( or difference mean for a Paired T test )}

\item{S}{Sample sd}

\item{CL}{Confidence Level ( default=0.95 )}

\item{N}{Population Size}

\item{t}{Mean to evaluate ( for Significance Tests )}
}
\description{
Inferences for One Sample T test ( or for a Paired T test ) are::
\cr Confidence Intervals
\cr Significance Tests
\cr
\cr Criteria for One Sample T test, ( Conditions for inference on a mean ) are:
\cr Randomness - A random sample or randomized experiment should be used to obtain the data.
\cr Normality - Sample size must be at least 30, unless population distribution is normal
\cr Independence - Sampling size n less then 10\% of the Population size
}
