% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XStatsZtest.R
\name{XStatsZtest}
\alias{XStatsZtest}
\title{Receives population_mean, observed_mean, observed_sd and sample size}
\usage{
XStatsZtest(pmean, omean, osd, n, alpha = 0.05)
}
\arguments{
\item{pmean}{Population Mean}

\item{omean}{Observed Mean}

\item{osd}{Observed sd}

\item{n}{Sample Size}

\item{alpha}{Significance Level ( probability of making a type I Error )}
}
\description{
Receives population_mean, observed_mean, observed_sd and sample size
}
\examples{
XSummary(iris$Sepal.Length)
}
