% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XStringPermutations.R
\name{XStringPermutations}
\alias{XStringPermutations}
\title{String Permutations}
\usage{
XStringPermutations(S)
}
\arguments{
\item{S}{Receives a String, Returns a Vector of Permutations}
}
\value{

}
\description{
String Permutations
}
\examples{
XStringPermutations("")
XStringPermutations(" ")
XStringPermutations("a")
XStringPermutations("aa")
XStringPermutations("ab")
XStringPermutations("abb")
XStringPermutations("abc")
XStringPermutations("a  b")
}
