% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XSummary.R
\name{XSummary}
\alias{XSummary}
\title{Enhanced Summary of a Vector}
\usage{
XSummary(d)
}
\arguments{
\item{d}{A Numeric Vector}
}
\value{

}
\description{
Enhanced Summary of a Vector
}
\examples{
XSummary(iris$Sepal.Length)
}
