% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XThresholdPlot.R
\name{XThresholdPlot}
\alias{XThresholdPlot}
\title{Computes the impact of choosing Threshold from 0 to 1 for a binary categorical variable}
\usage{
XThresholdPlot(predictions, reality)
}
\arguments{
\item{predictions}{Probability Vector with values from 0 to 1}

\item{reality}{Reality Vector contains the real 0 or 1 binary values}
}
\value{
optimalThreshold
}
\description{
Computes the impact of choosing Threshold from 0 to 1 for a binary categorical variable
}
