% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XBarPlot.R
\name{XBarPlot}
\alias{XBarPlot}
\title{BarPlot for one categorical variable (A table)}
\usage{
XBarPlot(data1)
}
\arguments{
\item{data1}{A Numeric Factor}
}
\value{

}
\description{
BarPlot for one categorical variable (A table)
}
\examples{
XBarPlot(as.factor(mtcars$cyl))
}
