% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XBarPlot2.R
\name{XBarPlot2}
\alias{XBarPlot2}
\title{BarPlot for one table (of 2 variables) - Contingency Tables}
\usage{
XBarPlot2(data1, data2)
}
\arguments{
\item{data1}{}

\item{data2}{}
}
\value{

}
\description{
BarPlot for one table (of 2 variables) - Contingency Tables
}
