% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XBarPlot3.R
\name{XBarPlot3}
\alias{XBarPlot3}
\title{BarPlot for one table (of 3 variables) - Contingency Tables}
\usage{
XBarPlot3(data1, data2, data3)
}
\arguments{
\item{data1}{}

\item{data2}{}

\item{data3}{}
}
\value{

}
\description{
BarPlot for one table (of 3 variables) - Contingency Tables
}
