% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XConfusionMatrix.R
\name{XConfusionMatrix}
\alias{XConfusionMatrix}
\title{Displays the Confusion Matrix from 2 vectors: predictions vs targets}
\usage{
XConfusionMatrix(predictions, targets)
}
\arguments{
\item{predictions}{}

\item{targets}{}
}
\description{
Displays the Confusion Matrix from 2 vectors: predictions vs targets
}
