% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XConfusionMatrix.R
\name{XConfusionMatrix}
\alias{XConfusionMatrix}
\title{Displays the Confusion Matrix from a table of 2x2}
\usage{
XConfusionMatrix(t)
}
\arguments{
\item{t}{}
}
\value{

}
\description{
Displays the Confusion Matrix from a table of 2x2
}
