% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XDescribeRandomVariable.R
\name{XDescribeRandomVariable}
\alias{XDescribeRandomVariable}
\title{Describe Random Variable}
\usage{
XDescribeRandomVariable(Xname, X, p)
}
\arguments{
\item{Xname}{A Name}

\item{X}{Vector of values}

\item{p}{Associated Vector of probabilities}
}
\value{

}
\description{
Describe Random Variable
}
