% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XEntropy.R
\name{XEntropy}
\alias{XEntropy}
\title{Calculate the Entropy of a Vector}
\usage{
XEntropy(v)
}
\arguments{
\item{v}{}
}
\value{
numeric
}
\description{
The entropy quantifies the expected value of the information contained in a vector
The Entropy is the average number of Yes/No questions we have to make
to guess which one object was picked
Entropy is Information
}
\examples{
XEntropy(c("1","2","3","4","5"))
}
