% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XLibrary.R
\name{XLibrary}
\alias{XLibrary}
\title{XLibrary}
\usage{
XLibrary(...)
}
\arguments{
\item{package}{0, 1 or more comma separated packages}
}
\description{
Receives a list of packages to Install/Load
Adds the received packages to a list of default packages to allways load
Installs missing packages
Load Packages
Inspect Packages for datasets
Creates 3 data frames on Global Environment:
Loaded.Packages
Loaded.Datasets
Loaded.Vignettes
}
\keyword{XLibrary}
