% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPlotChiSquare.R
\name{XPlotChiSquare}
\alias{XPlotChiSquare}
\title{The (non-central) Chi-Squared Distribution}
\usage{
XPlotChiSquare(df = 2, ConfidenceLevel = 0.95)
}
\arguments{
\item{df}{}

\item{ConfidenceLevel}{}
}
\description{
The mean on ChiSquare is = Degrees Of Freedom
SD is sqrt(2*DF)
}
