% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPlotPoisson.R
\name{XPlotPoisson}
\alias{XPlotPoisson}
\title{Plots The Poisson Distribution Poisson(lambda)}
\usage{
XPlotPoisson(lambda)
}
\arguments{
\item{L}{Number of events per unit time}
}
\description{
Plots The Poisson Distribution Poisson(lambda)
}
\examples{
XPlotPoisson(0.3)
}
