% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPowerSet.R
\name{XPowerSet}
\alias{XPowerSet}
\title{Returns a List with the Power Set of a set S}
\usage{
XPowerSet(v)
}
\arguments{
\item{v}{input vector}
}
\description{
The Power Set is the set of all subsets of S,
including the empty set and S itself
Cardinality of the Power Set must be 2^n with n = number of elements in the set
Remember that matematically, a Set cannot have repeated elements,
So this function will apply a unique() to the input vector
}
\examples{
x = XPowerSet()
x = XPowerSet(letters[1:1])
x = XPowerSet(letters[1:2])
x = XPowerSet(letters[1:3])
x = XPowerSet(c("bu","bu","cx","a"))
}
