% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XSearchFunction.R
\name{XSearchFunction}
\alias{XSearchFunction}
\title{Search which packages a function belongs to}
\usage{
XSearchFunction(functionName)
}
\description{
Searches in loaded environments
}
\examples{
\dontrun{
XSearchFunction(functionName)
}
}
