% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XStatsCIproportion.R
\name{XStatsCIproportion}
\alias{XStatsCIproportion}
\title{Estimates Confidence Interval for Population Proportion}
\usage{
XStatsCIproportion(n, p, CL = 0.95, N = 1e+06, t)
}
\arguments{
\item{n}{Sample Size}

\item{p}{Proportion Observed on the Sample (or proportion of the population for Significance Test)}

\item{CL}{Confidence Level ( default=0.95 )}

\item{N}{Population Size}

\item{t}{Proportion to evaluate ( Significance Test )}
}
\description{
Inferences on one proportion are: 
\cr Confidence Intervals 
\cr Significance Tests  
\cr
\cr Criteria for both, are:  
\cr Randomness
\cr Normality
\cr Independence
}
\examples{
XStatsCIproportion(100,0.54,0.95)
}
