% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XStatsTwoSampleTtest.R
\name{XStatsTwoSampleTtest}
\alias{XStatsTwoSampleTtest}
\title{Estimates Difference on the Means between two Simple Random Samples (SRS)}
\usage{
XStatsTwoSampleTtest(n1, m1, S1, n2, m2, S2, CL = 0.95, N = 1e+06)
}
\arguments{
\item{n1}{Sample 1 Size}

\item{m1}{Sample 1 mean}

\item{S1}{Sample 1 sd}

\item{n2}{Sample 2 Size}

\item{m2}{Sample 2 mean}

\item{S2}{Sample 2 sd}

\item{CL}{Confidence Level ( default=0.95 )}

\item{N}{Population Size}
}
\description{
Inferences for One Sample T test are:: 
\cr Confidence Intervals 
\cr Significance Tests 
\cr
\cr Criteria for One Sample T test, ( Conditions for inference on a mean ) are:  
\cr Randomness - A random sample or randomized experiment should be used to obtain the data.
\cr Normality - Sample size must be at least 30, unless population distribution is normal
\cr Independence - Sampling size n less then 10% of the Population size
}
\examples{
XStatsTwoSampleTtest(18,38.3,.27,24,38.9,.29,CL=.9)
}
